+++ packages/compiler/src/babel-helpers/ast.ts
@@ -56,41 +56,16 @@ export function isVineCompFnDecl(target: Node): boolean {
     target = target.declaration
   }
 
-  if (isFunctionDeclaration(target)) {
-    const returnStmt = target.body?.body.find(node => isReturnStatement(node))
-    if (!returnStmt) {
-      return false
-    }
-    return !!(returnStmt.argument && isVineTaggedTemplateString(returnStmt.argument))
+  if (isBabelFunctionTypes(target)) {
+    return checkFunctionReturnsVineTemplate(target)
   }
 
   if (isVariableDeclaration(target)) {
     const declValue = target.declarations[0].init
-    if (!declValue) {
+    if (!declValue || !isBabelFunctionTypes(declValue)) {
       return false
     }
-
-    if (isFunctionExpression(declValue)) {
-      const returnStmt = declValue.body?.body.find(node => isReturnStatement(node))
-      if (!returnStmt) {
-        return false
-      }
-      return !!(returnStmt.argument && isVineTaggedTemplateString(returnStmt.argument))
-    }
-
-    if (isArrowFunctionExpression(declValue)) {
-      const arrowFnReturns = (
-        isBlockStatement(declValue.body)
-          ? declValue.body.body.find(node => isReturnStatement(node))
-          : declValue.body
-      )
-      if (isReturnStatement(arrowFnReturns)) {
-        return !!(arrowFnReturns.argument && isVineTaggedTemplateString(arrowFnReturns.argument))
-      }
-      return isVineTaggedTemplateString(arrowFnReturns)
-    }
-
-    return false
+    return checkFunctionReturnsVineTemplate(declValue)
   }
 
   return false
@@ -161,6 +136,29 @@ export function isVineTaggedTemplateString(node: Node | null | undefined): node 
   }
   if (isVineVaporTaggedTemplateString(node)) {
     return true
+  }
+  return false
+}
+
+function checkFunctionReturnsVineTemplate(fn: BabelFunctionNodeTypes): boolean {
+  if (isFunctionDeclaration(fn) || isFunctionExpression(fn)) {
+    const returnStmt = fn.body?.body.find(node => isReturnStatement(node))
+    if (!returnStmt) {
+      return false
+    }
+    return !!(returnStmt.argument && isVineTaggedTemplateString(returnStmt.argument))
+  }
+
+  if (isArrowFunctionExpression(fn)) {
+    const arrowFnReturns = (
+      isBlockStatement(fn.body)
+        ? fn.body.body.find(node => isReturnStatement(node))
+        : fn.body
+    )
+    if (isReturnStatement(arrowFnReturns)) {
+      return !!(arrowFnReturns.argument && isVineTaggedTemplateString(arrowFnReturns.argument))
+    }
+    return isVineTaggedTemplateString(arrowFnReturns)
   }
 
   return false
@@ -287,6 +285,7 @@ export function getFunctionPickedInfos(fnDecl: Node): VineFnPickedInfo[] {
   }> = []
 
   let target = fnDecl
+  const isExportDefault = isExportDefaultDeclaration(fnDecl)
   if ((
     isExportNamedDeclaration(target)
     || isExportDefaultDeclaration(target)
@@ -305,7 +304,7 @@ export function getFunctionPickedInfos(fnDecl: Node): VineFnPickedInfo[] {
           && isIdentifier(decl.id)
         ) {
           pickedInfo.push({
-            fnDeclNode: fnDecl, // VariableDeclarator
+            fnDeclNode: fnDecl,
             fnItselfNode: decl.init,
             fnName: decl.id.name,
           })
@@ -318,11 +317,20 @@ export function getFunctionPickedInfos(fnDecl: Node): VineFnPickedInfo[] {
 
   if (isFunctionDeclaration(target)) {
     pickedInfo.push({
-      fnDeclNode: fnDecl, // FunctionDeclaration
+      fnDeclNode: fnDecl,
       fnItselfNode: target,
       fnName: target.id?.name ?? '',
     })
   }
+
+  if (isFunctionExpression(target) || isArrowFunctionExpression(target)) {
+    pickedInfo.push({
+      fnDeclNode: fnDecl,
+      fnItselfNode: target,
+      fnName: isExportDefault ? (isArrowFunctionExpression(target) ? 'default' : '') : '',
+    })
+  }
+
   return pickedInfo
 }