+++ packages/compiler/src/transform.ts
@@ -32,6 +32,93 @@ import {
 } from './transform/steps'
 
 /**
+ * Transform simple arrow functions without block statement body.
+ * e.g., `export default () => vine\`...\``
+ *
+ * These functions have no setup logic, only a template.
+ */
+function transformSimpleArrowFunction(
+  transformCtx: TransformContext,
+  vineCompFnCtx: import('./types').VineCompFnCtx,
+): void {
+  const { vineFileCtx, compilerHooks, inline, ssr, mergedImportsMap, templateComposer } = transformCtx
+  const ms = vineFileCtx.fileMagicCode
+  const isDev = compilerHooks.getCompilerCtx().options.envMode !== 'production'
+
+  const isExportDefaultAnonymous = vineCompFnCtx.isExportDefault && (vineCompFnCtx.fnName === 'default' || vineCompFnCtx.fnName === '')
+  const tempVarName = isExportDefaultAnonymous ? '__VineCompDefault' : vineCompFnCtx.fnName
+  const componentName = isExportDefaultAnonymous ? '__VineCompDefault' : vineCompFnCtx.fnName
+
+  // Import defineComponent
+  let vueImportsMeta = mergedImportsMap.get('vue')
+  if (!vueImportsMeta) {
+    vueImportsMeta = {
+      type: 'namedSpecifier',
+      specs: new Map(),
+    }
+    mergedImportsMap.set('vue', vueImportsMeta)
+  }
+  const vueImportsSpecs = (vueImportsMeta as any).specs
+  if (!vueImportsSpecs.has('defineComponent')) {
+    vueImportsSpecs.set('defineComponent', '_defineComponent')
+  }
+
+  // Compile template and get setup returns
+  const setupReturns = templateComposer.compileSetupFnReturns({
+    vineCompFnCtx,
+    vineFileCtx,
+    mergedImportsMap,
+  })
+
+  // Get render function from templateCompileResults
+  const renderFn = templateComposer.templateCompileResults.get(vineCompFnCtx) || ''
+
+  // Build component definition
+  const exportPrefix = isExportDefaultAnonymous ? '' : 'export '
+
+  const setupFn = inline
+    ? ''
+    : `setup(__props, { expose: __expose }) {
+      __expose();
+      return ${setupReturns};
+    }`
+
+  const renderFnAssignment = inline
+    ? ''
+    : `  ${renderFn}
+  __vine.${ssr ? 'ssrRender' : 'render'} = ${ssr ? '__sfc_ssr_render' : '__sfc_render'};
+`
+
+  const hmrIdAssignment = isDev
+    ? `  __vine.__hmrId = '${vineCompFnCtx.scopeId}';\n`
+    : ''
+
+  const hmrCreateRecord = (isDev && compilerHooks.getCompilerCtx().options.runtimeAdapter?.name !== 'rspack')
+    ? `\n\ntypeof __VUE_HMR_RUNTIME__ !== "undefined" && __VUE_HMR_RUNTIME__.createRecord(${tempVarName}.__hmrId, ${tempVarName});\n`
+    : ''
+
+  const defaultExport = vineCompFnCtx.isExportDefault
+    ? `\n\nexport default ${tempVarName};\n`
+    : ''
+
+  const generatedCode = `${exportPrefix}const ${tempVarName} = (() => {
+  const __vine = _defineComponent({
+    name: '${componentName}',
+    ${setupFn}
+  });
+  ${renderFnAssignment}
+  __vine.__vue_vine = true;
+  ${hmrIdAssignment}
+  return __vine;
+})();${hmrCreateRecord}${defaultExport}`
+
+  // Replace the entire function with generated component
+  const fnStart = vineCompFnCtx.fnDeclNode.start!
+  const fnEnd = vineCompFnCtx.fnDeclNode.end!
+  ms.overwrite(fnStart, fnEnd, generatedCode)
+}
+
+/**
  * Processing `.vine.ts` file transforming.
  *
  * Since we need to support sourcemap, we can't replace or cut-out too much code.
@@ -66,6 +153,7 @@ export function transformFile(
   for (const vineCompFnCtx of vineFileCtx.vineCompFns) {
     const vineCompFnBody = vineCompFnCtx.fnItselfNode?.body
     if (!isBlockStatement(vineCompFnBody)) {
+      transformSimpleArrowFunction(transformCtx, vineCompFnCtx)
       continue
     }