+++ packages/compiler/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vue-vine/compiler
 
+## 1.7.22
+
+### Patch Changes
+
+- Fix compiler for support simple annonymous vine arrow fn expr as default export.
+
+## 1.7.21
+
+### Patch Changes
+
+- Set rspack/rsbuild to official and fix missing VLS_elements variable declaration.
+
 ## 1.7.20
 
 ### Patch Changes